% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_o2pls.R
\name{calculate_sparse_vip}
\alias{calculate_sparse_vip}
\title{Calculate VIP scores for sparse PLS}
\usage{
calculate_sparse_vip(sparse_fit, X, Y, nc)
}
\arguments{
\item{sparse_fit}{Sparse PLS fit object}

\item{X}{Original X matrix}

\item{Y}{Original Y matrix}

\item{nc}{Number of components}
}
\description{
Calculate VIP scores for sparse PLS
}
\keyword{internal}
