% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{column_sums}
\alias{column_sums}
\title{get the total values for a matrix}
\usage{
column_sums(X)
}
\value{
sum value of the column
}
\description{
get the total values for a matrix
}
\keyword{internal}
