% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_plsda.R
\name{cv_sparse_o2pls_single}
\alias{cv_sparse_o2pls_single}
\title{Single cross-validation run for sparse O2PLS}
\usage{
cv_sparse_o2pls_single(
  X,
  Y,
  nc,
  nx,
  ny,
  keepX,
  keepY,
  validation,
  folds,
  measure
)
}
\description{
Single cross-validation run for sparse O2PLS
}
\keyword{internal}
