% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_plsda.R
\name{cv_sparse_plsda}
\alias{cv_sparse_plsda}
\title{Enhanced cross-validation for sparse PLS-DA}
\usage{
cv_sparse_plsda(X, Y, nc, keepX, validation, folds, dist)
}
\description{
Enhanced cross-validation for sparse PLS-DA
}
\keyword{internal}
