% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o2plscv.R
\name{.o2cv}
\alias{.o2cv}
\title{do cross-validation with group factors}
\usage{
.o2cv(X, Y, n, nx, ny, group, nr)
}
\arguments{
\item{X}{a Numeric matrix (input)}

\item{Y}{a Numeric matrix (input)}

\item{n}{Integer. Number of joint PLS components.}

\item{nx}{Integer. Number of orthogonal components in X}

\item{ny}{Integer. Number of orthogonal components in Y}

\item{group}{a vector to indicate the group for Y}

\item{nr}{Integer to indicate the folds for cross validation}
}
\value{
a list of o2pls results
}
\description{
do cross-validation with group factors
}
\author{
Kai Guo
}
\keyword{internal}
