% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.R
\name{.pls}
\alias{.pls}
\title{Internal PLS function}
\usage{
.pls(
  X,
  Y,
  nc,
  cv = TRUE,
  nr_folds = 5,
  tol = 1e-06,
  max_iter = 100,
  q2_threshold = 0.05
)
}
\description{
Internal function to perform PLS calculations
}
\keyword{internal}
