% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_plsda.R
\name{.sparse_o2cv}
\alias{.sparse_o2cv}
\title{Internal sparse cross-validation function}
\usage{
.sparse_o2cv(X, Y, nc, nx, ny, keepX, keepY, group, nr_folds)
}
\description{
Internal sparse cross-validation function
}
\keyword{internal}
