% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_o2pls.R
\name{evaluate_sparse_classification}
\alias{evaluate_sparse_classification}
\title{Evaluate sparse classification performance}
\usage{
evaluate_sparse_classification(sparse_fit, X, Y, dist)
}
\arguments{
\item{sparse_fit}{Sparse PLS fit object}

\item{X}{Original X matrix}

\item{Y}{Original Y labels}

\item{dist}{Distance metric}
}
\description{
Evaluate sparse classification performance
}
\keyword{internal}
