% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_o2pls.R
\name{extract_sparse_orthogonal_X}
\alias{extract_sparse_orthogonal_X}
\title{Extract orthogonal components with sparsity for X}
\usage{
extract_sparse_orthogonal_X(
  X_residual,
  T_joint,
  nx,
  keepX_orth,
  lambda_x,
  penalty
)
}
\arguments{
\item{X_residual}{Residual X matrix after joint component removal}

\item{T_joint}{Joint X scores}

\item{nx}{Number of X-orthogonal components}

\item{keepX_orth}{Vector of keepX values for orthogonal components}

\item{lambda_x}{L1 penalty parameter}

\item{penalty}{Penalty type}
}
\description{
Extract orthogonal components with sparsity for X
}
\keyword{internal}
