% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_o2pls.R
\name{find_optimal_sparse_params}
\alias{find_optimal_sparse_params}
\title{Find optimal sparse parameters using one standard error rule}
\usage{
find_optimal_sparse_params(error_rates, test_keepX)
}
\arguments{
\item{error_rates}{Matrix of error rates}

\item{test_keepX}{Vector of tested keepX values}
}
\description{
Find optimal sparse parameters using one standard error rule
}
\keyword{internal}
