% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_o2pls.R
\name{fit_sparse_pls}
\alias{fit_sparse_pls}
\title{Fit sparse PLS algorithm}
\usage{
fit_sparse_pls(X, Y, nc, keepX)
}
\arguments{
\item{X}{Predictor matrix}

\item{Y}{Response matrix}

\item{nc}{Number of components}

\item{keepX}{Vector of variables to keep per component}
}
\description{
Fit sparse PLS algorithm
}
\keyword{internal}
