% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMCCV_cpp}
\alias{getMCCV_cpp}
\title{MCCV sampling}
\usage{
getMCCV_cpp(x, n)
}
\value{
a vector with random sampling based on group-balanced Monte Carlo cross-validation
}
\description{
MCCV sampling
}
\keyword{internal}
