% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_o2pls.R
\name{\%||\%}
\alias{\%||\%}
\title{Helper function to check if object is null or use default}
\usage{
x \%||\% y
}
\description{
Helper function to check if object is null or use default
}
\keyword{internal}
