% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{loadings.o2plsda}
\alias{loadings.o2plsda}
\title{extract the loading value from the O2PLSDA analysis}
\usage{
\method{loadings}{o2plsda}(x, loading = "Xloading", ...)
}
\arguments{
\item{loading}{}
}
\description{
extract the loading value from the O2PLSDA analysis
}
