% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{loadings}
\alias{loadings}
\title{Extract the loadings from an O2PLS fit}
\usage{
loadings(x, ...)
}
\arguments{
\item{x}{Object of class \code{O2pls}}

\item{...}{For consistency}
}
\value{
Loading matrix
}
\description{
This function extracts loading parameters from an O2PLS fit
}
