% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{O2pls-class}
\alias{O2pls-class}
\title{Class "O2pls"
This class represents the Annotation information}
\description{
Class "O2pls"
This class represents the Annotation information
}
\section{Slots}{

\describe{
\item{\code{X}}{a Numeric matrix (input)}

\item{\code{Y}}{a Numeric matrix (input)}

\item{\code{params}}{paramaters ysed in o2pls analysis}

\item{\code{results}}{list of o2pls results}
}}

\author{
Kai Guo
}
\keyword{classes}
