% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{opls}
\alias{opls}
\title{orthogonal scores algorithn of partial leat squares (opls) projection}
\usage{
opls(X, Y, ncomp, maxiter, tol)
}
\value{
a list of opls result
}
\description{
orthogonal scores algorithn of partial leat squares (opls) projection
}
\keyword{internal}
