% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Score, vip or loading plot for the O2PLS or O2PLSDAresults}
\usage{
plot(x, ...)
}
\arguments{
\item{x}{an O2pls object}

\item{...}{Other arguments}
}
\value{
a ggplot2 object
}
\description{
Score, vip or loading plot for the O2PLS or O2PLSDAresults
}
\author{
Kai Guo
}
