% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sparse_plsda_classification}
\alias{plot_sparse_plsda_classification}
\title{Plot Sparse PLS-DA classification boundaries}
\usage{
plot_sparse_plsda_classification(x, component, group, ...)
}
\description{
Plot Sparse PLS-DA classification boundaries
}
\keyword{internal}
