% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sparse_plsda_loadings}
\alias{plot_sparse_plsda_loadings}
\title{Plot Sparse PLS-DA loadings}
\usage{
plot_sparse_plsda_loadings(x, component, top, ...)
}
\description{
Plot Sparse PLS-DA loadings
}
\keyword{internal}
