% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sparse_plsda_scores}
\alias{plot_sparse_plsda_scores}
\title{Plot Sparse PLS-DA scores}
\usage{
plot_sparse_plsda_scores(x, component, group, color, ellipse, ...)
}
\description{
Plot Sparse PLS-DA scores
}
\keyword{internal}
