% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sparse_plsda_selection}
\alias{plot_sparse_plsda_selection}
\title{Plot Sparse PLS-DA variable selection}
\usage{
plot_sparse_plsda_selection(x, ...)
}
\description{
Plot Sparse PLS-DA variable selection
}
\keyword{internal}
