% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sparse_plsda_vip}
\alias{plot_sparse_plsda_vip}
\title{Plot Sparse PLS-DA VIP scores}
\usage{
plot_sparse_plsda_vip(x, component, top, ...)
}
\description{
Plot Sparse PLS-DA VIP scores
}
\keyword{internal}
