% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sparse_s4_biplot}
\alias{plot_sparse_s4_biplot}
\title{Plot SparseO2pls biplot (S4 implementation)}
\usage{
plot_sparse_s4_biplot(x, component, group, top, color, title, ...)
}
\description{
Plot SparseO2pls biplot (S4 implementation)
}
\keyword{internal}
