% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sparse_s4_comparison}
\alias{plot_sparse_s4_comparison}
\title{Plot SparseO2pls comparison with regular O2PLS (S4 implementation)}
\usage{
plot_sparse_s4_comparison(x, title, ...)
}
\description{
Plot SparseO2pls comparison with regular O2PLS (S4 implementation)
}
\keyword{internal}
