% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sparse_s4_contribution}
\alias{plot_sparse_s4_contribution}
\title{Plot SparseO2pls variable contribution (S4 implementation)}
\usage{
plot_sparse_s4_contribution(x, component, block, top, title, ...)
}
\description{
Plot SparseO2pls variable contribution (S4 implementation)
}
\keyword{internal}
