% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sparse_s4_diagnostic}
\alias{plot_sparse_s4_diagnostic}
\title{Plot SparseO2pls diagnostic plots (S4 implementation)}
\usage{
plot_sparse_s4_diagnostic(x, title, ...)
}
\description{
Plot SparseO2pls diagnostic plots (S4 implementation)
}
\keyword{internal}
