% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sparse_s4_loadings}
\alias{plot_sparse_s4_loadings}
\title{Plot SparseO2pls loadings (S4 implementation)}
\usage{
plot_sparse_s4_loadings(x, component, block, top, title, ...)
}
\description{
Plot SparseO2pls loadings (S4 implementation)
}
\keyword{internal}
