% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sparse_s4_scores}
\alias{plot_sparse_s4_scores}
\title{Plot SparseO2pls scores (S4 implementation)}
\usage{
plot_sparse_s4_scores(x, component, block, group, color, ellipse, title, ...)
}
\description{
Plot SparseO2pls scores (S4 implementation)
}
\keyword{internal}
