% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sparse_s4_selection}
\alias{plot_sparse_s4_selection}
\title{Plot SparseO2pls variable selection summary (S4 implementation)}
\usage{
plot_sparse_s4_selection(x, title, ...)
}
\description{
Plot SparseO2pls variable selection summary (S4 implementation)
}
\keyword{internal}
