% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sparse_s4_sparsity}
\alias{plot_sparse_s4_sparsity}
\title{Plot SparseO2pls sparsity pattern (S4 implementation)}
\usage{
plot_sparse_s4_sparsity(x, block, title, ...)
}
\description{
Plot SparseO2pls sparsity pattern (S4 implementation)
}
\keyword{internal}
