% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_stability_barplot_impl}
\alias{plot_stability_barplot_impl}
\title{Implementation of stability barplot}
\usage{
plot_stability_barplot_impl(x, component, top_n, ...)
}
\description{
Implementation of stability barplot
}
\keyword{internal}
