% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_stability_heatmap_impl}
\alias{plot_stability_heatmap_impl}
\title{Implementation of stability heatmap}
\usage{
plot_stability_heatmap_impl(x, component, top_n, ...)
}
\description{
Implementation of stability heatmap
}
\keyword{internal}
