% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_stability_summary_impl}
\alias{plot_stability_summary_impl}
\title{Implementation of stability summary}
\usage{
plot_stability_summary_impl(x, ...)
}
\description{
Implementation of stability summary
}
\keyword{internal}
