% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tuning_heatmap_impl}
\alias{plot_tuning_heatmap_impl}
\title{Implementation of tuning heatmap}
\usage{
plot_tuning_heatmap_impl(x, metric, ...)
}
\description{
Implementation of tuning heatmap
}
\keyword{internal}
