% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tuning_line_impl}
\alias{plot_tuning_line_impl}
\title{Implementation of tuning line plot}
\usage{
plot_tuning_line_impl(x, metric, ...)
}
\description{
Implementation of tuning line plot
}
\keyword{internal}
