% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_plsda.R
\name{predict.sparse_plsda}
\alias{predict.sparse_plsda}
\title{Enhanced predict method for sparse PLS-DA}
\usage{
predict.sparse_plsda(object, newdata, dist = "max.dist", ...)
}
\description{
Enhanced predict method for sparse PLS-DA
}
