% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_plsda.R
\name{predict.SparseO2pls}
\alias{predict.SparseO2pls}
\title{Enhanced predict method for sparse O2PLS}
\usage{
predict.SparseO2pls(object, newdata, ...)
}
\description{
Enhanced predict method for sparse O2PLS
}
