% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_o2pls.R
\name{predict_o2pls_classification}
\alias{predict_o2pls_classification}
\title{Predict O2PLS for classification}
\usage{
predict_o2pls_classification(o2_fit, X_test, Y_train)
}
\arguments{
\item{o2_fit}{O2PLS model object}

\item{X_test}{Test data matrix}

\item{Y_train}{Training labels for centroid calculation}
}
\description{
Predict O2PLS for classification
}
\keyword{internal}
