% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_o2pls.R
\name{predict_o2pls_regression}
\alias{predict_o2pls_regression}
\title{Predict O2PLS for regression}
\usage{
predict_o2pls_regression(o2_fit, X_test)
}
\arguments{
\item{o2_fit}{O2PLS model object}

\item{X_test}{Test data matrix}
}
\description{
Predict O2PLS for regression
}
\keyword{internal}
