% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print.O2pls}
\alias{print.O2pls}
\title{Print the summary of O2PLS results.}
\usage{
\method{print}{O2pls}(fit, ...)
}
\arguments{
\item{x}{An O2pls object}
}
\description{
Print the summary of O2PLS results.
}
\author{
Kai Guo
}
