% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print.plsda}
\alias{print.plsda}
\title{Print the summary of plsda results.}
\usage{
\method{print}{plsda}(x, ...)
}
\arguments{
\item{x}{An plsda object}

\item{...}{For consistency}
}
\description{
Print the summary of plsda results.
}
\examples{
X <- matrix(rnorm(500),10,50)
Y <- rep(c("a","b"),each=5)
fit <- plsda(X,Y,2)
print(fit)
}
\author{
Kai Guo
}
