% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{scores.O2pls}
\alias{scores.O2pls}
\title{Extract the scores from an O2PLS fit}
\usage{
\method{scores}{O2pls}(x, score = c("Xjoint", "Yjoint", "Xorth", "Yorth"), ...)
}
\arguments{
\item{x}{Object of class \code{O2pls}}

\item{score}{the scores matrix for one of "Xjoint", "Yjoint", "Xorth", "Yorth"}

\item{...}{Other arguments}
}
\value{
score matrix
}
\description{
This function extracts scores parameters from an O2PLS fit
}
