% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{scores.o2plsda}
\alias{scores.o2plsda}
\title{Extract the scores from an O2PLS DA analysis}
\usage{
\method{scores}{o2plsda}(x, ...)
}
\arguments{
\item{x}{Object of class \code{o2plsda}}

\item{...}{Other arguments}
}
\value{
score matrix
}
\description{
Extract the scores from an O2PLS DA analysis
}
\author{
Kai Guo
}
