% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{scores.plsda}
\alias{scores.plsda}
\title{Extract the scores PLSDA analysis}
\usage{
\method{scores}{plsda}(x, ...)
}
\arguments{
\item{x}{Object of class \code{plsda}}

\item{...}{Other arguments}
}
\value{
score matrix
}
\description{
Extract the scores PLSDA analysis
}
\author{
Kai Guo
}
