% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{scores}
\alias{scores}
\title{Extract the scores from an O2PLS fit}
\usage{
scores(x, ...)
}
\arguments{
\item{x}{Object of class \code{O2pls}}

\item{...}{For consistency}
}
\value{
Scores matrix
}
\description{
This function extracts score matrices from an O2PLS fit
}
