% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{selected_var_names}
\alias{selected_var_names}
\title{Get names of selected variables}
\usage{
selected_var_names(x, type = "X", ...)
}
\arguments{
\item{x}{A sparse model object}

\item{type}{Character. Which variables to extract: "X", "Y", or "both"}

\item{...}{Additional arguments passed to selected_vars}
}
\value{
Character vector of variable names or list of name vectors
}
\description{
Get names of selected variables
}
