% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{selected_vars.default}
\alias{selected_vars.default}
\title{Default method for selected_vars}
\usage{
\method{selected_vars}{default}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
Default method for selected_vars
}
