% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{selected_vars.sparse_plsda}
\alias{selected_vars.sparse_plsda}
\title{Extract selected variables from sparse PLS-DA models}
\usage{
\method{selected_vars}{sparse_plsda}(x, ...)
}
\arguments{
\item{x}{A sparse_plsda object}

\item{...}{Additional arguments (currently unused)}
}
\value{
Vector of selected variable indices
}
\description{
Extract selected variables from sparse PLS-DA models
}
