% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SparseO2pls-class}
\alias{SparseO2pls-class}
\title{Class "SparseO2pls"
This class represents sparse O2PLS analysis results}
\description{
Class "SparseO2pls"
This class represents sparse O2PLS analysis results
}
\section{Slots}{

\describe{
\item{\code{X}}{Numeric matrix (input X)}

\item{\code{Y}}{Numeric matrix (input Y)}

\item{\code{params}}{List of parameters including sparsity}

\item{\code{results}}{List of sparse O2PLS results}

\item{\code{sparsity}}{List of sparsity-related information}

\item{\code{preprocessing}}{list of preprocessing information}

\item{\code{call}}{The matched call used to generate the object}
}}

\author{
Kai Guo
}
