% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{sparsity_info}
\alias{sparsity_info}
\title{Get sparsity information from sparse models}
\usage{
sparsity_info(x)
}
\arguments{
\item{x}{A sparse model object}

\item{...}{Additional arguments}
}
\value{
List containing sparsity statistics
}
\description{
Get sparsity information from sparse models
}
