% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{summary.O2pls}
\alias{summary.O2pls}
\title{Summary of an O2PLS fit}
\usage{
\method{summary}{O2pls}(fit)
}
\arguments{
\item{fit}{a O2pls object}
}
\value{
Detail of O2PLS results
}
\description{
Summary of an O2PLS fit
}
\author{
Kai Guo
}
