% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{summary.plsda}
\alias{summary.plsda}
\title{Summary of an plsda object}
\usage{
\method{summary}{plsda}(object, ...)
}
\arguments{
\item{object}{a plsda object}

\item{...}{For consistency}
}
\value{
Detail of plsda results
}
\description{
Summary of an plsda object
}
\examples{
X <- matrix(rnorm(500),10,50)
Y <- rep(c("a","b"),each=5)
fit <- plsda(X,Y,2)
summary(fit)
}
\author{
Kai Guo
}
