% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_plsda.R
\name{tune_sparse_o2pls}
\alias{tune_sparse_o2pls}
\title{Cross-validation for Sparse O2PLS}
\usage{
tune_sparse_o2pls(
  X,
  Y,
  nc_range = 1:3,
  nx_range = 0:2,
  ny_range = 0:2,
  keepX_range = seq(10, 50, 10),
  keepY_range = seq(10, 50, 10),
  validation = "Mfold",
  folds = 5,
  measure = "RMSE"
)
}
\arguments{
\item{X}{Numeric matrix (samples x variables)}

\item{Y}{Numeric matrix (samples x variables)}

\item{nc_range}{Range of joint components to test}

\item{nx_range}{Range of X-orthogonal components to test}

\item{ny_range}{Range of Y-orthogonal components to test}

\item{keepX_range}{Range of keepX values to test}

\item{keepY_range}{Range of keepY values to test}

\item{validation}{Validation method}

\item{folds}{Number of folds}

\item{measure}{Performance measure}
}
\value{
TuneResult object with optimal parameters
}
\description{
Performs cross-validation to evaluate sparse O2PLS models
}
