% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_o2pls.R
\name{validate_tuning_inputs}
\alias{validate_tuning_inputs}
\title{Validate inputs for O2PLS parameter tuning}
\usage{
validate_tuning_inputs(X, Y, nc_range, nx_range, ny_range, validation, measure)
}
\arguments{
\item{X}{Numeric matrix}

\item{Y}{Numeric matrix or factor}

\item{nc_range}{Vector of joint component numbers}

\item{nx_range}{Vector of X-orthogonal component numbers}

\item{ny_range}{Vector of Y-orthogonal component numbers}

\item{validation}{Validation method}

\item{measure}{Performance measure}
}
\description{
Validate inputs for O2PLS parameter tuning
}
\keyword{internal}
