% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{vip}
\alias{vip}
\title{Extract the VIP values from the O2PLS-DA object}
\usage{
vip(x)
}
\arguments{
\item{x}{the o2plsda object or plsda object}
}
\value{
a data frame
}
\description{
Extract the VIP values from the O2PLS-DA object
}
