% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.R
\name{.pls}
\alias{.pls}
\title{Partial least squares discriminant analysis}
\usage{
.pls(X, Y, nc, cv = TRUE, nr_folds = 5)
}
\arguments{
\item{X}{a matrix of predictor variables.}

\item{Y}{a single vector indicate the group}

\item{nc}{the number of pls components .}

\item{cv}{logical indicating whether cross-validation will be performed or not (suggest TRUE).}

\item{nr_folds}{nr_folds Integer to indicate the folds for cross validation.
@return list with PLS results}
}
\description{
Perform a PLS discriminant analysis
}
\keyword{internal}
