% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{loadings.plsda}
\alias{loadings.plsda}
\title{extract the loading value from the PLSDA analysis}
\usage{
\method{loadings}{plsda}(x, ...)
}
\arguments{
\item{x}{Object of class \code{plsda}}

\item{...}{For consistency}
}
\description{
extract the loading value from the PLSDA analysis
}
