% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{loadings}
\alias{loadings}
\alias{loadings.O2pls}
\title{Extract the loadings from an O2PLS fit}
\usage{
loadings(x, ...)

\method{loadings}{O2pls}(x, loading = c("Xjoint", "Yjoint", "Xorth", "Yorth"), ...)
}
\arguments{
\item{x}{Object of class \code{O2pls}}

\item{...}{For consistency}

\item{loading}{the loadings for one of "Xjoint", "Yjoint", "Xorth", "Yorth"}
}
\value{
Loading matrix

Loading matrix
}
\description{
This function extracts loading parameters from an O2PLS fit

This function extracts loading parameters from an O2PLS fit
}
