% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o2plscv.R
\name{o2cv}
\alias{o2cv}
\title{Cross validation for O2PLS}
\usage{
o2cv(
  X,
  Y,
  nc,
  nx,
  ny,
  group = NULL,
  nr_folds,
  ncores = 1,
  scale = FALSE,
  center = FALSE
)
}
\arguments{
\item{X}{a Numeric matrix (input)}

\item{Y}{a Numeric matrix (input)}

\item{nc}{Integer. Number of joint PLS components.}

\item{nx}{Integer. Number of orthogonal components in X}

\item{ny}{Integer. Number of orthogonal components in Y}

\item{group}{a vector to indicate the group for Y}

\item{nr_folds}{Integer to indicate the folds for cross validation}

\item{ncores}{Integer. Number of CPUs to use for cross validation}

\item{scale}{boolean values determining if data should be scaled or not}

\item{center}{boolean values determining if data should be centered or not}
}
\description{
Cross validation for O2PLS
}
\author{
Kai Guo
}
