% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print.O2pls}
\alias{print.O2pls}
\title{Print the summary of O2PLS results.}
\usage{
\method{print}{O2pls}(x, ...)
}
\arguments{
\item{x}{An O2pls object}

\item{...}{For consistency}
}
\description{
Print the summary of O2PLS results.
}
\examples{
X <- matrix(rnorm(50),10,5)
Y <- matrix(rnorm(50),10,5)
object <- o2pls(X,Y,1,1,1)
print(object)
}
\author{
Kai Guo
}
